Test for specific Exception messages
Because we're not using custom exception types to denote when different types
of exceptions occur, we can assert against the message to be sure the
exact exception being thrown is the one we expect. Else in some cases a
exception of the expected type could be thrown earlier in the method and
the test would still pass.